UP: utils: Add parameter video_src
Update paramaters now that --camera and --webcam are no longer necessary
since now it is possible to automatically detect which device are being
used. The parameter video_src was added so a video file or a video
device can be selected as video source.
Update multimedia/utils to add video source with the new parameter.

Signed-off-by: Alifer Moraes <alifer.moraes@nxp.com>