UP: apps: Update Switch Image application
The switch image application used to run using files that was already
part of NXP BSP, which implies errors if those files are removed or
modified. Now those files were storaged in an server and are downloaded
when the application runs for the first time.

Other than that, improvements were made to make the code simpler and
standardized.

Signed-off-by: Alifer Moraes <alifer.moraes@nxp.com>